+++ packages/hierarchy-builder/src/hierarchy-builder/internal/Common.ts
@@ -3,18 +3,14 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import { Observable } from "rxjs";
 import { assert } from "@itwin/core-bentley";
 import {
   BaseGroupingParams,
   GroupingParams,
-  HierarchyNode,
   HierarchyNodeKey,
   InstanceHierarchyNodeProcessingParams,
   InstancesNodeKey,
-  ParentHierarchyNode,
   ProcessedCustomHierarchyNode,
-  ProcessedHierarchyNode,
   ProcessedInstanceHierarchyNode,
 } from "../HierarchyNode";
 import { getLogger } from "../Logging";
@@ -168,30 +164,4 @@ export function createOperatorLoggingNamespace(operatorName: string) {
 export function julianToDateTime(julianDate: number): Date {
   const millis = (julianDate - 2440587.5) * 86400000;
   return new Date(millis);
-}
-
-/** @internal */
-export interface ChildNodesObservables {
-  processedNodes: Observable<ProcessedHierarchyNode>;
-  finalizedNodes: Observable<HierarchyNode>;
-  hasNodes: Observable<boolean>;
-}
-
-/** @internal */
-export class ChildNodesCache {
-  private _map = new Map<string, ChildNodesObservables>();
-
-  private createKey(node: ParentHierarchyNode | undefined): string {
-    return node ? `${JSON.stringify(node.parentKeys)}+${JSON.stringify(node.key)}` : "";
-  }
-
-  public add(parentNode: ParentHierarchyNode | undefined, value: ChildNodesObservables) {
-    const key = this.createKey(parentNode);
-    assert(!this._map.has(key));
-    this._map.set(key, value);
-  }
-
-  public get(parentNode: ParentHierarchyNode | undefined): ChildNodesObservables | undefined {
-    return this._map.get(this.createKey(parentNode));
-  }
 }
